+++ examples/authentication/app.rb
@@ -1,5 +1,5 @@
 require 'pry'
-require 'nylas'
+require 'nylas-legacy'
 require 'sinatra'
 require 'omniauth'
 require 'omniauth-google-oauth2'
@@ -24,7 +24,7 @@ end
   send(method, "/auth/:provider/callback") do
     auth_hash = env['omniauth.auth'] # => OmniAuth::AuthHash
 
-    api = Nylas::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'])
+    api = NylasLegacy::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'])
     nylas_token = api.authenticate(name: auth_hash[:info][:name], email_address: auth_hash[:info][:email],
                                    provider: :gmail,
                                    settings: { google_client_id: ENV['GOOGLE_CLIENT_ID'],