+++ examples/helpers.rb
@@ -1,5 +1,5 @@
 $LOAD_PATH.unshift(File.expand_path(File.join(File.dirname(__FILE__), "..", "lib")))
-require "nylas"
+require "nylas-legacy"
 require "faker"
 require "pry"
 require "awesome_print"