+++ examples/plain-ruby/contacts.rb
@@ -1,7 +1,7 @@
 require_relative '../helpers'
 # An executable specification that demonstrates how to use the Nylas Ruby SDK to interact with the API. It
 # follows the rough structure of the [Nylas API Reference](https://docs.nylas.com/reference).
-api = Nylas::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'],
+api = NylasLegacy::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'],
                      access_token: ENV['NYLAS_ACCESS_TOKEN'])
 # Retrieving a count of contacts
 demonstrate { api.contacts.count }
@@ -66,7 +66,7 @@ contact_with_picture = api.contacts.where(email: 'contact-with-picture@example.c
 demonstrate { contact_with_picture.picture_url }
 begin
 demonstrate { contact_with_picture.picture }
-rescue Nylas::ResourceNotFound
+rescue NylasLegacy::ResourceNotFound
   puts "That contacts picture didn't propagate yet :/"
 end