+++ examples/plain-ruby/deltas.rb
@@ -3,7 +3,7 @@ require_relative '../helpers'
 # An executable specification that demonstrates how to use the Nylas Ruby SDK to interact with the Nylas #
 # Deltas API.
 # See https://docs.nylas.com/reference#deltas
-api = Nylas::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'],
+api = NylasLegacy::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'],
                      access_token: ENV['NYLAS_ACCESS_TOKEN'])
 
 
@@ -27,7 +27,7 @@ demonstrate { deltas_from_cursor.find_each.map(&:id).count }
 # 5 delta's
 demonstrate { deltas_from_cursor.take(5).map(&:to_h) }
 
-# Models are cast to Nylas::Model objects
+# Models are cast to NylasLegacy::Model objects
 demonstrate { deltas_from_cursor.first&.model&.class }
 
 # And can be viewed directly