+++ examples/plain-ruby/threads.rb
@@ -2,7 +2,7 @@ require_relative '../helpers'
 
 # An executable specification that demonstrates how to use the Nylas Ruby SDK to interact with the API. It
 # follows the rough structure of the [Nylas API Reference](https://docs.nylas.com/reference).
-api = Nylas::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'],
+api = NylasLegacy::API.new(app_id: ENV['NYLAS_APP_ID'], app_secret: ENV['NYLAS_APP_SECRET'],
                      access_token: ENV['NYLAS_ACCESS_TOKEN'])
 
 
@@ -25,7 +25,7 @@ demonstrate { { starred: reloaded_thread.starred, unread: reloaded_thread.unread
 demonstrate do
   begin
     api.threads.create
-  rescue Nylas::ModelNotCreatableError => e
+  rescue NylasLegacy::ModelNotCreatableError => e
     "#{e.class}: #{e.message}"
   end
 end
@@ -36,7 +36,7 @@ end
 demonstrate do
   begin
     thread.destroy
-  rescue Nylas::ModelNotDestroyableError => e
+  rescue NylasLegacy::ModelNotDestroyableError => e
     "#{e.class}: #{e.message}"
   end
 end