+++ lib/nylas-legacy/physical_address.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module NylasLegacy
+  # Structure to represent the Physical Address schema
+  # @see https://docs.nylas.com/reference#contactsid
+  class PhysicalAddress
+    include Model::Attributable
+    attribute :format, :string
+    attribute :type, :string
+    attribute :street_address, :string
+    attribute :postal_code, :string
+    attribute :state, :string
+    attribute :city, :string
+    attribute :country, :string
+  end
+end