+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@
 require "simplecov"
 SimpleCov.start
 
-require "nylas"
+require "nylas-legacy"
 require "pry"
 require "webmock/rspec"
 require "rspec-json_matcher"
@@ -23,7 +23,7 @@ end
 # Illustrates all the types and such a model can be built out of. Used for testing the generic Model
 # functionality without conflating it with actual Models
 class FullModel
-  include Nylas::Model
+  include NylasLegacy::Model
   self.creatable = true
   self.showable = true
   self.listable = true
@@ -48,7 +48,7 @@ class FullModel
 end
 
 class NotCreatableModel
-  include Nylas::Model
+  include NylasLegacy::Model
   self.resources_path = "/not_creatable_collection"
 
   attribute :id, :string
@@ -66,7 +66,7 @@ class NotCreatableModel
 end
 
 class NotUpdatableModel
-  include Nylas::Model
+  include NylasLegacy::Model
   self.resources_path = "/not_updatable_collection"
 
   attribute :id, :string
@@ -84,7 +84,7 @@ class NotUpdatableModel
 end
 
 class NonFilterableModel < FullModel
-  include Nylas::Model
+  include NylasLegacy::Model
   self.resources_path = "/non_filterable_collection"
 
   attribute :id, :string